fix : remove duplicate board view
